+++ ExonSurfer/specificity/blast.py
@@ -13,7 +13,7 @@ from ExonSurfer.resources import resources
 #                    blast module FUNCTION DEFINITION SITE                    #
 ###############################################################################
             
-def run_blast_list(fastaf, out, db_path, species,
+def run_blast_list(fastaf, out, db_path, species, tomerge = True, 
                    blastn_path = "blastn", 
                    num_threads = 4):
     """
@@ -44,18 +44,18 @@ def run_blast_list(fastaf, out, db_path, species,
     # remove transcript version information
     df["subject id"] = df["subject id"].map(lambda x: x.split(".")[0])
     
-    
-    table = pd.read_csv(os.path.join(resources.get_blastdb_path(species),
-                                     resources.IDS_TABEL), 
-                        sep = "\t", names=["id", "gene"], header=None)
-    # remove transcript version information
-    table["id"] = table["id"].map(lambda x: x.split(".")[0])
-    
-    df = pd.merge(df, table, left_on="subject id", right_on = "id", how = "left")
-    
-    # overwrite annotated blast result
-    df.to_csv(out, sep = "\t", index = False)
-    os.remove(out)
+    if tomerge == True: 
+        table = pd.read_csv(os.path.join(resources.get_blastdb_path(species),
+                                         resources.IDS_TABEL), 
+                            sep = "\t", names=["id", "gene"], header=None)
+        # remove transcript version information
+        table["id"] = table["id"].map(lambda x: x.split(".")[0])
+        
+        df = pd.merge(df, table, left_on="subject id", right_on = "id", how = "left")
+        
+        # overwrite annotated blast result
+        df.to_csv(out, sep = "\t", index = False)
+        os.remove(out)
     
     return df
 
@@ -153,7 +153,21 @@ def filter_big_blast(blast_df, design_df, maxrows = 15000):
 ###############################################################################
 
 def check_row_spec(blast_df, row, design_df, max_sep, t_gene, t_transcripts): 
-    
+    """ USED WITH cDNA BLAST DATABASE
+    This function is applied to every row in the blast DF. It extracts the 
+    alignment and checks if there are other alignments (from the same primer 
+    pair) that could cause off-target amplification and annotates these as: 
+        - detected: every transcript from the same gene amplified by the primers
+        - other_transcripts: transcripts different than the target ones
+        - other_genes: every transcript from different genes
+    Args: 
+        blast_df      [in] (pd.df)  Blast alignments df
+        row           [in] (df row) Row (from blast df) that will be analyzed 
+        design_df     [in] (pd.df)  DF with primer information
+        max_sep       [in] (int)    Max sep between als to take as off-target
+        t_gene        [in] (str)    Target gene name (gene symbol)
+        t_transcripts [in] (l|str)  List fo transcripts or "ALL"
+    """
     # get pair name
     pair_name = row["query id"][:-2]
     
@@ -180,7 +194,7 @@ def check_row_spec(blast_df, row, design_df, max_sep, t_gene, t_transcripts):
                             design_df.loc[pair_name, "other_transcripts"] += row["subject id"] +";"
                     else: 
                         design_df.loc[pair_name, "other_genes"] += row["subject id"] +";"
-        
+                        
 ###############################################################################
 
 def check_specificity(blast_df, design_df, t_gene, t_transcript, max_sep): 
@@ -207,22 +221,88 @@ def check_specificity(blast_df, design_df, t_gene, t_transcript, max_sep):
 
 ###############################################################################
 
-def show_ot_for_pair(transcripts, other_genes, other_transcripts): 
+def check_row_genomic(blast_df, row, design_df, max_sep): 
+    """ USED WITH GENOMIC BLAST DATABASE
+    This function is applied to every row in the blast DF. It extracts the 
+    alignment and checks if there are other alignments (from the same primer 
+    pair) that could cause off-target amplification and annotates these as 
+    genomic_amp (chromosome, start and end position). 
+    Args: 
+        blast_df      [in] (pd.df)  Blast alignments df
+        row           [in] (df row) Row (from blast df) that will be analyzed 
+        design_df     [in] (pd.df)  DF with primer information
+        max_sep       [in] (int)    Max sep between als to take as off-target
+    """    
+    # get pair name
+    pair_name = row["query id"][:-2]
+    
+    # iterate blast df WITHOUT row being inspected
+    ppair_num = (row["query id"][:-2] + "_5", 
+                 row["query id"][:-2] + "_3")
+    
+    poss_offts = blast_df.index[(blast_df["query id"].isin(ppair_num)) &\
+                                (blast_df["subject id"] == row["subject id"])]
+
+    for i in [ind for ind in poss_offts if ind != row.name]: 
+        # if opposite strand
+        if row["strand"] != blast_df.iloc[i]["strand"]: 
+            # if close enough
+            if abs(row["s. start"] - blast_df.iloc[i]["s. start"]) < max_sep: 
+                
+                # get start and end positions
+                spos = min(row["s. start"], row["s. end"], 
+                           blast_df.iloc[i]["s. start"], blast_df.iloc[i]["s. end"])
+                epos = max(row["s. start"], row["s. end"], 
+                           blast_df.iloc[i]["s. start"], blast_df.iloc[i]["s. end"]) 
+                
+                # build chromosome location
+                chrom_loc = "({}:{}-{})".format(row["subject id"], spos, epos)
+                
+                # if not already annotated
+                if chrom_loc not in design_df.loc[pair_name, "genomic_amp"]: 
+                    design_df.loc[pair_name, "genomic_amp"] += chrom_loc +";"
+
+###############################################################################
+
+def check_genomic_specificity(blast_df, design_df, max_sep): 
+    """ USED WITH GENOMIC BLAST DATABASE
+    This function annotates genomic_amp (chromosome, start and end position) in
+    the final dataframe. 
+    Args: 
+        blast_df      [in] (pd.df)  Blast alignments df
+        row           [in] (df row) Row (from blast df) that will be analyzed 
+        design_df     [in] (pd.df)  DF with primer information
+        max_sep       [in] (int)    Max sep between als to take as off-target
+    """     
+    # new columns to fill in 
+    design_df["genomic_amp"] = ""
+
+    # check off-targets
+    blast_df.apply(lambda row: check_row_genomic(blast_df, row, design_df, 
+                                                 max_sep), 
+                   axis = 1)    
+        
+    return design_df
+
+###############################################################################
+
+def show_ot_for_pair(transcripts, other_genes, other_transcripts, genomic_amp): 
     """
     This function says if a primer pair has off-targets to return or not. 
     Args: 
         transcripts [in] (str|l)      List of transcripts or ALL
         other_genes [in] (str)        List of other genes or empty
         other_genes [in] (str)        List of other transcripts or empty
+        genomic_amp [in] (str)        List of possible genomic amplification
         to_return [out] (int)         1 if there are off_targets, 0 if not
     """
     to_return = 0 #  default not show 
     
     if transcripts == "ALL": 
-        if other_genes != "": 
+        if other_genes != "" or genomic_amp != "": 
             to_return = 1
     else: 
-        if other_genes != "" or other_transcripts != "":
+        if other_genes != "" or other_transcripts != "" or genomic_amp != "":
             to_return = 1
             
     return to_return
@@ -238,7 +318,8 @@ def show_off_targets(final_df, transcripts):
     """
     final_df["off_targets"] = final_df.apply(lambda row: show_ot_for_pair(transcripts, 
                                                                           row["other_genes"], 
-                                                                          row["other_transcripts"]), 
+                                                                          row["other_transcripts"], 
+                                                                          row["genomic_amp"]), 
                                              axis = 1)
     
     return final_df
